refactor: get related course
